Move RequestDetails out of BinaryDownloadHandler.
